refactor: modularize directive handler factories
